+++ lys_em/__init__.py
@@ -1,5 +1,6 @@
 from .FFT import fft, ifft
 from .tem import TEM
+from .space import FunctionSpace 
 from .scatteringFactor import scatteringFactor
 from .kinematical import debyeWallerFactors, formFactors, structureFactors
 from .crystalPotential import CrystalPotential
