+++ test/test_multislice.py
@@ -3,10 +3,10 @@ import numpy as np
 from numpy.testing import assert_array_almost_equal
 
 from lys_mat import CrystalStructure,Atom
-from lys_em import TEM
+from lys_em import TEM, FunctionSpace
 from lys_em.consts import m, e, h, hbar
 from lys_em.scatteringFactor import projectedPotential
-from lys_em.multislice import FunctionSpace, calcMultiSliceDiffraction, _apply
+from lys_em.multislice import calcMultiSliceDiffraction, _apply
 from lys_em.crystalPotential import _Slices, _Potentials
 
 
@@ -20,7 +20,7 @@ class MultiSlice_test(unittest.TestCase):
         self.Au_single = CrystalStructure([20, 20, 20, 90, 90, 90], [Atom("Au", (0, 0, 0))])
 
     def test_FunctionSpace(self):
-        sp = FunctionSpace(self.Au, 10, 10)
+        sp = FunctionSpace.fromCrystal(self.Au, 10, 10)
         ref = np.linalg.norm(self.Au.InverseLatticeVectors()[0])
 
         assert_array_almost_equal(sp.kvec[0, 0], [0, 0])
@@ -28,7 +28,7 @@ class MultiSlice_test(unittest.TestCase):
         assert_array_almost_equal(sp.kvec[1, 0], [ref, 0])
         assert_array_almost_equal(sp.kvec[2, 0], [ref * 2, 0])
 
-        sp = FunctionSpace(self.Au_single, 512 * 2, 512 * 2, division=1)
+        sp = FunctionSpace.fromCrystal(self.Au_single, 512 * 2, 512 * 2)
 
     def test_TaTe2(self):
         # Compare result with pre-calculated results at 2023/12/14.
@@ -42,15 +42,15 @@ class MultiSlice_test(unittest.TestCase):
 
     def test_Propagation(self):
         crys = CrystalStructure(self.Au.cell, [])
-        sp = FunctionSpace(crys, 128, 128, 10)
+        sp = FunctionSpace.fromCrystal(crys, 128, 128)
         b = TEM(60e3)
-        V_rs = _Slices(crys, sp).getPotentialTerms(b)
+        V_rs = _Slices(crys, sp, division=10).getPotentialTerms(b)
         pot = _Potentials(V_rs, sp.kvec, crys.unit[2][0], crys.unit[2][1], 1)
 
         phi = np.zeros((128, 128))
         phi[0][0] = 1
 
-        P_k = sp.getPropagationTerm(b.wavelength)
+        P_k = sp.getPropagationTerm(b.wavelength, crys.c/10)
         phi = _apply(phi, pot, P_k*sp.mask)
 
         # theoretical solution (propagation after a in free space)
@@ -85,21 +85,21 @@ class CrystalPotential_test(unittest.TestCase):
 
     def test_Slices(self):
         # check slicing
-        sp = FunctionSpace(self.Au, 10, 10, division=9)
+        sp = FunctionSpace.fromCrystal(self.Au, 10, 10)
         b = TEM(60e3, 0)
-        slices = _Slices(self.Au, sp)
+        slices = _Slices(self.Au, sp, division=9)
         n = np.sum([len(s.atoms) for s in slices._slices])
         self.assertEqual(n, len(self.Au.atoms))
 
-        sp = FunctionSpace(self.TaTe2, 10, 10, division=9)
-        slices = _Slices(self.TaTe2, sp)
+        sp = FunctionSpace.fromCrystal(self.TaTe2, 10, 10)
+        slices = _Slices(self.TaTe2, sp, division=9)
         n = np.sum([len(s.atoms) for s in slices._slices])
         self.assertEqual(n, len(self.TaTe2.atoms))
 
         # check single atom potential
         N = 1024  # if this value is too small, this test fails because of window function in Function Space.
-        sp = FunctionSpace(self.Au_single, N, N, division=1)
-        V_k = _Slices(self.Au_single, sp)._calculatePotential(self.Au_single) * b.wavelength * b.relativisticMass / m  # A^2
+        sp = FunctionSpace.fromCrystal(self.Au_single, N, N)
+        V_k = _Slices(self.Au_single, sp, division=1)._calculatePotential(self.Au_single) * b.wavelength * b.relativisticMass / m  # A^2
         V_r = sp.IFT(V_k * sp.mask)
 
         r = np.linspace(0, self.Au_single.a, N, endpoint=False)
@@ -114,8 +114,8 @@ class CrystalPotential_test(unittest.TestCase):
         self.assertAlmostEqual(V_r[0][2], sf[2])
 
         # check potential calculation
-        sp = FunctionSpace(self.Au, 10, 10, division=1)
-        slices = _Slices(self.Au, sp)
+        sp = FunctionSpace.fromCrystal(self.Au, 10, 10)
+        slices = _Slices(self.Au, sp, division=1)
         pot = slices._calculatePotential(slices._slices[0])
         self.assertTrue(np.abs((pot[0, 0] / pot[0, 2]) - 1.725) < 1e-3)
         self.assertTrue(np.abs(pot[0, 1]) < 1e-5)
